[refactor] Rearrange code
Rearrange to better match standard go project layouts
