+++ internal/announcers/types.go
