+++ internal/stage/fake.go
