+++ internal/stage/minio/types.go
