+++ internal/stage/s3/types.go
