+++ internal/validators/kafka/kafka_suite_test.go
@@ -1,0 +1,15 @@
+package kafka_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+	l "github.com/redhatinsights/insights-ingress-go/internal/logger"
+)
+
+func TestKafka(t *testing.T) {
+	RegisterFailHandler(Fail)
+	l.InitLogger()
+	RunSpecs(t, "Kafka Suite")
+}
