+++ internal/version/version_suite_test.go
@@ -1,0 +1,16 @@
+package version_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	l "github.com/redhatinsights/insights-ingress-go/internal/logger"
+)
+
+func TestInventory(t *testing.T) {
+	RegisterFailHandler(Fail)
+	l.InitLogger()
+	RunSpecs(t, "Version Suite")
+}
