--- track/track_suite_test.go
@@ -1,15 +1,0 @@
-package track_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-	l "github.com/redhatinsights/insights-ingress-go/logger"
-)
-
-func TestTrack(t *testing.T) {
-	RegisterFailHandler(Fail)
-	l.InitLogger()
-	RunSpecs(t, "Track Suite")
-}
